requirements.txt
absl-py==0.12.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.23
filelock==3.0.12
future==0.18.2
fvcore==0.1.5.post20210423
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.10
iopath==0.1.8
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.9
Pillow==8.2.0
portalocker==2.3.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
RandAugment @ git+https://github.com/ildoonet/pytorch-randaugment@48b8f509c4bbda93bbe733d98b3fd052b6e4c8ae
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
six==1.16.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
timm @ git+https://github.com/rwightman/pytorch-image-models.git@5f9aff395c224492e9e44248b15f44b5cc095d9c
tokenizers==0.9.4
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
transformers==4.2.0
typing-extensions==3.10.0.0
urllib3==1.26.4
VQAtools @ git+https://github.com/seo-95/VQAtools.git@5ce906daac69f4c0ed771bb39e8f7b0757767337
warmup-scheduler @ git+https://github.com/seo-95/pytorch-gradual-warmup-lr.git@99ce139a003d117b256fe72b0d5f33f0597f771a
Werkzeug==1.0.1
yacs==0.1.8
